pyproject.toml
[tool.poetry]
name = "mediaurlembedfixerbot"
version = "0.1.0"
description = ""
authors = ["Matheus Albino <matheusalb3213@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^0.21.1"
python-telegram-bot = "^20.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
